# Remove base type check to accommodate nested types such as ROW(MAP)